requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.6
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
Flask-Testing==0.8.1
PyMySQL==1.1.1
hijri-converter==2.3.1
python-dateutil==2.9.0.post0
cryptography==43.0.3
python-dotenv==1.0.0
pytest==8.0.0
pylint==3.0.3
black==24.1.1
coverage==7.4.1
Flask-Mail==0.9.1
Flask-Babel==4.0.0
Babel==2.14.0
pytz==2024.1
sentry-sdk[flask]==1.40.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
redis==5.0.1
gunicorn==21.2.0
supervisor==4.2.5
pytest-cov==4.1.0
pytest-flask==1.3.0
mypy==1.8.0
bandit==1.7.7
Flask-DebugToolbar==0.14.1
WTForms==3.0.1
email_validator==2.1.1
requests==2.31.0
